Improve url handling
`urllib.request.urlretrieve` is legacy interface it might be deprecated
in the future.
https://docs.python.org/3/library/urllib.request.html#urllib.request.urlretrieve

If downloaded fits files need `hp.ud_grade` download it to a temp file
following example:
https://docs.python.org/3/howto/urllib2.html#fetching-urls
The temp file will be removed when file closed.
This could avoid writing to disk twice.

If download directly to the disk, use simple file object operation.